move creating new section to a dom-util function
