+++ src/converter/modules/blocks/create-shade-box.js
@@ -1,4 +1,4 @@
-import { toBlock } from '../utils/dom-utils.js';
+import { toBlock, createNewSectionForBlock } from '../utils/dom-utils.js';
 
 /**
  *
@@ -7,20 +7,7 @@ import { toBlock } from '../utils/dom-utils.js';
 export default function createShadeBox(document) {
   Array.from(document.querySelectorAll('.sp-wrapper')).forEach(
     (shadeBoxElement) => {
-      const section = shadeBoxElement.parentElement;
-      const shadeBoxSection = document.createElement('div');
-      const subsequenSection = document.createElement('div');
-      section.after(shadeBoxSection);
-
-      let nextPointer = shadeBoxElement.nextElementSibling;
-      if (nextPointer) {
-        shadeBoxSection.after(subsequenSection);
-        while (nextPointer) {
-          const next = nextPointer.nextElementSibling;
-          subsequenSection.append(nextPointer);
-          nextPointer = next;
-        }
-      }
+      const shadeBoxSection = createNewSectionForBlock(shadeBoxElement);
 
       shadeBoxSection.append(...shadeBoxElement.children);
       shadeBoxSection.append(
