+++ src/converter/modules/utils/dom-utils.js
@@ -147,6 +147,33 @@ const getHeadingLevel = (element) => {
 };
 
 /**
+ * Creates a new section for the given element/block. The resulting dom contains up to 3 sections: 
+ * 1) the original section the block was in, containing elements before the block
+ * 2) the new section containing only the block
+ * 3) If there are elements after the block, a new section containing those elements
+ * @param {Element} block the block element
+ * @returns the new section the block is contained in
+ */
+export const createNewSectionForBlock = (block) => {
+  const section = block.parentElement;
+  const blockSection = document.createElement('div');
+  const subsequenSection = document.createElement('div');
+  section.after(blockSection);
+
+  let nextPointer = block.nextElementSibling;
+  if (nextPointer) {
+    section.after(subsequenSection);
+    while (nextPointer) {
+      const next = nextPointer.nextElementSibling;
+      subsequenSection.append(nextPointer);
+      nextPointer = next;
+    }
+  }
+
+  return blockSection;
+};
+
+/**
  *
  * @param {Document} document
  */
